Optimize fields
